requirements.txt
PyMuPDF==1.21.1
Flask==2.1.0
flask_cors==3.0.10
flask_restful==0.3.9
nltk==3.7
requests==2.28.1
resume-parser==0.8.4
session_info==1.0.0
spacy==2.3.9
werkzeug==2.0.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
requests-oauthlib==1.3.1
importlib-metadata==3.2.0
psycopg2-binary
pandas==1.5.3
pdfkit==1.0.0
mammoth==1.5.0
python-dotenv
gunicorn==20.1.0
locationtagger==0.0.1
elastic-apm[flask]
blinker
Authlib==1.2.0
